mito-ai: notification on app repload
